Replaced math.h double functions by their float equivalent
Where applicable. In hopes it speeds up things.

+++ kernels/volk/volk_32f_cos_32f.h
@@ -158,7 +158,7 @@ static inline void
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = cos(*aPtr++);
+    *bPtr++ = cosf(*aPtr++);
   }
 }
 
@@ -240,7 +240,7 @@ volk_32f_cos_32f_u_sse4_1(float* bVector, const float* aVector, unsigned int num
 
   number = quarterPoints * 4;
   for(;number < num_points; number++){
-    *bPtr++ = cos(*aPtr++);
+    *bPtr++ = cosf(*aPtr++);
   }
 }
 
@@ -315,7 +315,7 @@ volk_32f_cos_32f_generic(float* bVector, const float* aVector, unsigned int num_
   unsigned int number = 0;
 
   for(; number < num_points; number++){
-    *bPtr++ = cos(*aPtr++);
+    *bPtr++ = cosf(*aPtr++);
   }
 }
